requirements.txt
# requirements.txt
flask
searoute
python-dotenv